% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregateModels}
\alias{aggregateModels}
\title{Aggregate flowReMix model fits}
\usage{
aggregateModels(x, verbose = TRUE, summarizeCoefs = FALSE)
}
\arguments{
\item{x}{A \code{list} of flowReMix model fits or a \code{character} vector of \code{rds} files storing model fits.}

\item{verbose}{\code{logical}. Display a progress bar.}
}
\description{
Average flowReMix model fits based on multiple starting points.
}
