% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degreeFromStringFun}
\alias{degreeFromStringFun}
\title{degreeFromStringFun
A function to parse the degree of a subset from its name.}
\usage{
degreeFromStringFun(x, split = "\\\\+")
}
\arguments{
\item{x}{\code{character} cell subset name from \code{getSubsets}}
}
\value{
\code{numeric} vector of degree of functionality.
}
\description{
degreeFromStringFun
A function to parse the degree of a subset from its name.
}
\details{
Assumes that a cell subset name follows a specific pattern where
the positive functions are simply listed in the name. For example, "8+/A+B+C+" would be  a degree 3
subset of CD8+ cells.
}
\examples{
data(fit505)
degreeFromStringFun(getSubsets(fit505))

}
\seealso{
\code{\link{flowReMixPFS}} \code{\link{weightForPFS}}
}
