% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetResponseRcpp.R
\name{flowReMix}
\alias{flowReMix}
\title{Fitting a Mixture of Mixed Effect Models for Binomial Data}
\usage{
flowReMix(formula, subject_id, cell_type = NULL, cluster_variable,
  data = parent.frame(), cluster_assignment = NULL, weights = NULL,
  covariance = c("sparse", "dense", "diagonal"), ising_model = c("sparse",
  "dense", "none"), regression_method = c("betabinom", "binom", "sparse",
  "robust"), iterations = 10, parallel = TRUE, verbose = TRUE,
  control = NULL)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}. The response
should be a matrix of two column matrix with first column containing the
counts of the cell subsets of interest and the second column the difference
between the reference count and the cell count.}

\item{subject_id}{a vector identifying the subjects.}

\item{cell_type}{a factor vector identifying which cell type each row in the
data set refers to.}

\item{cluster_variable}{a variable with respect to which clustering will be
done. See description for more detail.}

\item{data}{a data frame containing the variables in the model. It is
advisable to include the \code{subject_id}, \code{cell_type} and
\code{cluster_variable} variables in the data frame.}

\item{cluster_assignment}{an optional matrix of known cluster assignments.
Must include all subject/cell_type combinations. See description for more
detail.}

\item{weights}{an option vector of weights.}

\item{covariance}{the method to be used for estimating the covariance
structure of the random effects. \code{\link[PDSCE]{pdsoft.cv}} will be
used by default.}

\item{ising_model}{a method for estimating the Ising model.
\code{\link[IsingFit]{IsingFit}} will be used by default.}

\item{regression_method}{the regression method to be used. Default option is
the \code{\link[stats]{glm}} function with family = "binomial".}

\item{iterations}{the number of stochastic-EM itreations to perform.}

\item{parallel}{\code{logical}. Use parallel processing to fit the model. Default TRUE.}

\item{verbose}{whether to print information regrading the fitting process as
the optimization algorithm runs.}

\item{control}{an optional object of \code{\link{flowReMix_control}} class.}
}
\value{
\code{flowReMix} returns an object of class \code{flowReMix} which
contains the following variables:
\itemize{
\item \code{coefficients} a list, each component of which is a vector of
regession coefficients corresponding to a single cell type.
\item \code{posteriors} a matrix containing the posterior probabilities for
response computed for each subject/cell-type combination.
\item \code{levelProbs} a vector of the marginal estimated probabilities of
response estiamted for each cell subset.
\item \code{randomEffects} the estimated random effects for each
subject/cell-type.
\item \code{covariance} the estimated covariance structure for the random
effects.
\item \code{isingCov} the estimated \code{covariance} structure of the ising model.
\item \code{isingfit} if \code{ising_model = "sparse"} then the object returned
by the \code{\link[IsingFit]{IsingFit}} function. \code{NULL} otherwise.
\item \code{dispersion} the over-dispersion estimated for each cell-subset. If
regression method is not "betabinomial" then this will be a vector of large
constants.
\item \code{assignmentList} a list of matrices containing the posterior cluster
assignemnt sampled for each subject at the last iteration of the stochastic
EM algorithm.
\item \code{data} the input data frame.
\item \code{subject_id} the value of the subject_id argument used in the call.
}
}
\description{
\code{flowReMix} fits a mixture of mixed effect models
to binomial or over-dispersed binomial data. The package was specifically
designed for analyzing flow-cytometry cell-count data but may be suitable
for other purposes as well.
}
\details{
flowReMix fits a mixture of mixed effects regression models for
binomial data. Accordingly, the response supplied in the \code{formula}
must contain be a two column matrix the first column of which is the number
of successes and the second column is the number of failiures. In the
context of flow-cytomery count the left column would be the cell counts and
the right columns the parent counts minus the cell count. The right side of
the formula should include any number of fixed effects. For details on how
the function processes the formula object see, for example, the
documentation for the \code{\link[stats]{glm}} function.

The model fit by the function is a hierchical one, assuming the existence
of subjects and one or more cell-types for each subject. the
\code{subject_id} variable identifies different rows in the dataset as
corresponding to measurements taken from specific subjects. The model
assumes the existence of a random intercept for each \code{cell_type}.

The \code{cluster_variable} identifies which variable out of the covariates
corresponds to the variable with respect to which clustering should be
performed. The model assumes that the effect of cluster variable (and
corresponding interactions) are either always zero or non-zero. For
flow-cytomery experiments the \code{cluster_variable} will typically be an
indicator for whether the stimulation introduced into the blood sample was
an antigen or a control. A response status (zero or non-zero) is estimated
for each subject/cell-type combination. The dependence between the
cell-subsets is modeled via an Ising model.

\code{cluster_assignment} is an optional variable which allows the user to
pre-specificy some known cluster assignments. For example, in vaccine
studies we could expect all subjects who received a placebo treatment to be
non-responders across all cell-subsets. This variable should be three
column matrix, the first column of which should contain all unique values
of \code{subject_id}, the second should column should contain all unique
values of \code{cell_type} and in total the matrix should include all
\code{subject_id} and \code{cell_type} combinations. The third column is an
integer which takes the value 0 if the cell-type/subject combination is
non-response, 1 if it is response and -1 if the response status is unknown
and must be estimated.

The fitting algorithm uses one of three methods for estimating the
covariance structure of the random effects. A diagonal covariance structure
will be estimated if \code{covariance = "diagonal"}. A dense covariance
structure will be estimated with no penalization will be estimated if
\code{covariance = "dense"}. This may produce a singual covariance
structure if the number of subjects is smaller than the number of
cell-types. A sparse covariance matrix is estimated via the
\code{\link[PDSCE]{pdsoft.cv}} function by default.

The ising model describing the dependence between the response/non-resposne
status of the different cell-types can be estimated via three methods. If
\code{ising_model} is set to \code{"none"} then an independence model is
assumed. If the \code{ising_model} is set to \code{"dense"} then the ising
model is estimated via a set of firth regressions
(\code{\link[logistf]{logistf}}), one for each node in the graph. The
default option is \code{"sparse"}, where the
\code{\link[IsingFit]{IsingFit}} method is used.

\code{regression_method} specifies which function should be used for
estimating the reqression coefficients conditionally on the values of the
random effects and cluster assignments. If the default option
\code{"binom"} is chosen then a binomial model is fit using the
\code{\link[stats]{glm}} function. Otherwise, if \code{"betabinom"} option
is selected then a beta-binomial regression model is estimated with the
\code{\link[gamlss]{gamlss}} function. We recommend using the
\code{"sparse"} method which uses the \code{\link[glmnet]{cv.glmnet}}
procedure if the number of subjects is small and the number of predictors
is large.
}
