% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reMixMisc.R
\name{flowReMix_control}
\alias{flowReMix_control}
\title{Auxiliary for Controlling flowReMix Fitting.}
\usage{
flowReMix_control(updateLag = 5, randomAssignProb = 0, nsamp = 20,
  lastSample = NULL, initMHcoef = 0.4, nPosteriors = NULL,
  maxDispersion = 10^3, minDispersion = 10^8, isingInit = -4,
  keepEach = 5, centerCovariance = TRUE, intSampSize = 100,
  initMethod = NULL, ncores = NULL, preAssignCoefs = 0,
  markovChainEM = TRUE)
}
\arguments{
\item{updateLag}{number of iterations before the algorithm is assumed to
converge, at which time the parameter estimates will be aggregated.}

\item{randomAssignProb}{an optional parameter for adding noise to the cluster
assignments generated by the Gibbs sampler. Should only be changed if ising
model estimates are unstable.}

\item{nsamp}{number of Gibbs/componentwise MH cycles to perform for each
subject. Must be larger than keepEach.}

\item{initMHcoef}{the initial value for the shrinkage/inflation to perform on
the estimated covariance in the componentwise MH sampler. The initial value
does not matter much as this parameter self-tunes as the algorithm runs and
usually converges to a good value within a few iterations.}

\item{nPosteriors}{number of posterior samples to take per subject. If left
as \code{NULL} then a value will be determined within flowReMix according
to a preset formula.}

\item{maxDispersion}{the maximum overdispersion level allowed. The lower the
value of the variable the more overdispersion is allowed. Must be larger
than 0.}

\item{keepEach}{one out of how many Gibbs/MH samples to keep. This is used to
reduce the dependence between posterior samples.}

\item{centerCovariance}{whether to center random effect estimates before
computing the covariance estimate or not.}

\item{intSampSize}{number of importance samples to take when performing
univariate numerical integration in the Gibbs sampler.}

\item{initMethod}{the method used to initialize the regression coefficients.
Options are either "sparse" for \code{\link[glmnet]{cv.glmnet}} or "binom"
for \code{\link[stats]{glm}}. If left as \code{NULL} then it will be determined
according to the regression_method specified for the \code{flowReMix} call.}
}
\value{
An object of type \code{flowReMix_control}.
}
\description{
Auxiliary function for \code{\link{flowReMix}} fitting. Can be
  used to generate an appropriate object for modifying the fitting process of
  the stochastic EM algorithm used by \code{flowReMix}.
}
