% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FlowReMixFromCOMPASSContainers}
\alias{FlowReMixFromCOMPASSContainers}
\title{Create a table suitable for FlowReMix modeling from a list of COMPASS container objects}
\usage{
FlowReMixFromCOMPASSContainers(x)
}
\arguments{
\item{x}{\code{list} of COMPASSContainer objects}
}
\value{
\code{data.frame} suitable for input to flowReMix.
}
\description{
Create a table suitable for FlowReMix modeling from a list of COMPASS container objects
}
\examples{
\dontrun{
#x is a list of COMPASSContainer
FlowReMixFromCOMPASSContainers(x)

}
}
