% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flowReMixPFS}
\alias{flowReMixPFS}
\title{perSubsetPFS
Compute the PFS (polyfunctionality score) per subject and stimulation group}
\usage{
flowReMixPFS(x, M, stimVar = NULL, parentVar = NULL, outcomeVar = NULL,
  ...)
}
\arguments{
\item{x}{\code{flowReMix} model object}

\item{M}{\code{numeric} the max possible cell subset degree, the number of functions measured.}

\item{stimVar}{\code{name} Unquoted  name of the stimulation variable in the data. e.g. stim}

\item{parentVar}{\code{name} Unquoted name of the parent cell population variable in the data. e.g. parent}
}
\value{
\code{data.frame} of polyfunctionality scores for each cell subset and subject, weighted appropriately.
}
\description{
perSubsetPFS
Compute the PFS (polyfunctionality score) per subject and stimulation group
}
\details{
Requires that the data table has a variable for stimulation and for cell population parent.

The user can pass in a function to the argument  \code{parser=}. This function parses a cell subset name string  and
returns the degree of functionality for the cell subset. The default is \code{\link{degreeFromStringFun}}
}
\note{
Uses rlang quosures to pass variable information. PFS are NOT normalized to the total number of possible subsets (n*(n+1)/2), like in COMPASS.
}
\examples{
data(fit505)
flowReMixPFS(fit505,M=5, stimVar = stimGroup, parentVar = parent)
flowReMixPFS(fit505,M=5,stimVar=stimGroup,parentVar=parent,outcomeVar=infection)
}
\seealso{
\code{\link{degreeFromStringFun}} \code{\link{weightForPFS}}
}
