% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIsing}
\alias{getIsing}
\title{getIsing}
\usage{
getIsing(x)
}
\arguments{
\item{x}{\code{flowReMix} model fit.}
}
\value{
\code{isingStability} slot from the model.
}
\description{
getIsing
}
\examples{
data(fit505)
getIsing(fit505)
}
