% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPosteriors}
\alias{getPosteriors}
\title{getPosteriors}
\usage{
getPosteriors(x)
}
\arguments{
\item{x}{\code{flowReMix} model fit}
}
\value{
\code{data.frame} of posterior probabilities for each subject and subset in the model.
}
\description{
getPosteriors
}
\examples{
data(fit505)
getPosteriors(fit505)

}
