% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSubsets}
\alias{getSubsets}
\title{getSubsets}
\usage{
getSubsets(x)
}
\arguments{
\item{x}{\code{flowReMix} object}
}
\value{
\code{vector} of subset names fitted in the model.
}
\description{
getSubsets
}
\examples{
data(fit505)
getSubsets(fit505)

}
