% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmerMixture.R
\name{glmmMixture}
\alias{glmmMixture}
\title{Estimate a Mixture of Regressions with Univariate Random Effect}
\usage{
glmmMixture(formula, N = NULL, id, data = parent.frame(), treatment,
  weights = NULL, maxiter = 100, tol = 0.001, nAGQ = 1)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{N}{a vector of total/parent population counts.}

\item{id}{id variable to identify observations that should share the same random effect.}

\item{treatment}{which is the treatment variable? Same variable must appear in the formula.}

\item{weights}{prior weights for the data.}

\item{maxiter}{maximum number of iterations of the optimization algorithm.}

\item{tol}{tolerance - used to decide when to stop the optimization loop.}

\item{nAGQ}{see documentation of lme4::glmer.}
}
\description{
Fits a regression mixture to data with a univariate random effect.
}
