% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCindex.R
\name{hcIndex}
\alias{hcIndex}
\title{A function for computing the Higher-Criticism index}
\usage{
hcIndex(fit, groups = "all", target, direction = c("two.sided", "less",
  "greater"), a0 = 0.75, nPermutations = 1000, nBootstrap = 200,
  verbose = TRUE)
}
\arguments{
\item{fit}{a object of class 'flowReMix'}

\item{groups}{the groups for which the HC index should be computed. By
default, the function constructs an HC index based on all cell-subsets.}

\item{target}{a factor variable with two levels for which a classifier
will be constructed}

\item{a0}{Only the top a0% of cell-subsets will be considered for inclusion
in the HC index}
}
\description{
Computes the Higher-Critcism index, computes a p-value via a permutation
and conducts model averaging through bootstrap.
}
