% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parser}
\alias{parser}
\title{Cell subset label parser
Parses cell subset labels into matrix of functions.}
\usage{
parser(x, separator, functions)
}
\arguments{
\item{x}{\code{data.frame0} input with a column corresponding to the cell subset function labels}

\item{separator}{\code{string} to split the label on}

\item{functions}{\code{name} bare, unquoted name of the column holding the function label}
}
\value{
matrix of individual functions.
}
\description{
Best explained via the example.
}
\examples{
dat = data.frame(name=c("IFNg+TNFa+IL2+","IFNg+","IL2+","CD154+TNF+"),stringsAsFactors=FALSE)
parser(dat,separator = "\\\\+",functions=name)
}
\seealso{
\code{\link{plotChordDiagram}}
}
