% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{plot}
\alias{plot}
\alias{plot.flowReMix}
\title{plot a flowReMix fit object}
\usage{
\method{plot}{flowReMix}(x,...)
}
\arguments{
\item{x}{The model fit of class \code{flowReMix} returned by the fitting function.}

\item{...}{additional arguments, see description.}
}
\description{
Generate a plot of a flowReMix model fit. Various plot types are supported.
\itemize{
\item{subject}{The name of the subject variable, as an unquoted variable.}
\item{target}{unquoted variable name in the data corresponding to the outcome variable. Required.}
\item{varname}{character string used to label the outcome variable on the plot legend.}
\item{type}{character specifying the plot type. can be "scatter", "boxplot", "FDR", "ROC", "ising". Required.}
}
}
