% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotChordDiagram}
\alias{plotChordDiagram}
\title{plotChordDiagram
Generate a chord diagram from a flowReMix fit.}
\usage{
plotChordDiagram(x, threshold = 0.6, varnames = c("stim", "parent",
  "functions"), separator = "/", function.separator = "\\\\+",
  track.order = c("functions", "parent", "stim"),
  sort.order = c("functions", "parent", "stim"), inner.track = "parent",
  track.height = 2, track.margin = c(0, 0), functions.name = "functions",
  wrap.at = c("\\\\+", "", ""), label.cex = c(0.5, 0.5, 0.5),
  facing = "clockwise", parsefun = NULL)
}
\arguments{
\item{x}{\code{flowReMix} fit}

\item{threshold}{\code{numeric} threshold for the minimum edge weight in the graph (exclusive).}

\item{varnames}{\code{character} variable names to define by splitting subsets. e.g. \code{c("parent","stim","function")}}

\item{separator}{\code{character} separator to split subset string into variables.}

\item{function.separator}{\code{character} separator to split multifunction string  into individual functions.}

\item{track.order}{\code{character} vector of the same length as \code{varnames} specifying the order of the variables in the tracks.}

\item{sort.order}{\code{character} vector of the same length as \code{varnames} specifying the order of the variables in the tracks.}

\item{inner.track}{\code{character} specifying which variable in \code{varnames} is on the inner track.}

\item{track.height}{\code{numeric} vector of track heights, length 1 or \code{length(varnames)}}

\item{track.margin}{\code{numeric} vector of length 2 specifying the track margin.}

\item{functions.name}{\code{character} length 1, name of the column holding the functions.}

\item{label.cex}{\code{numeric} scaling factors for the labels in each track.Should be same length as \code{varnames}.}

\item{facing}{\code{character} passed to circos.text. \code{facing=c("inside", "outside", "reverse.clockwise", "clockwise",
"downward", "bending", "bending.inside", "bending.outside")}}

\item{parsefun}{\code{function} that parses multiple function labels to single function matrix.}

\item{wrap.width}{\code{numeric} string width to wrap annotation labels for all tracks.}
}
\description{
plotChordDiagram
Generate a chord diagram from a flowReMix fit.
}
\details{
Parses the \code{subsets} labels into factors that can be used to annotate tracks on a chord diagram.
Connections correspond to edges in the ising stability graph with a minumum \code{threshold} propensity.
A label such as "env/8+/IL2+" with an argument of \code{separator="/"} and \code{varnames=c("stim","parent","functions")} would
cause the cell subset label to be parsed into three variables corresponding to the stimulation, the
parent cell population, and the cytokine function.
}
\examples{
data(fit505)
#also defined in the package namespace
parser = function(x,separator,functions){
  functions=enquo(functions)
   dat = x\%>\% mutate(func = strsplit(!!functions, split = separator)) \%>\%
 tidyr::unnest()\%>\%tidyr::spread(func,func)
 dat[is.na(dat)]=""
 dat
}
plotChordDiagram(fit505,
   threshold = 0.6,
   varnames=c("stim","parent","functions"),
   track.order = c("functions","stim","parent"),
   sort.order=c("stim","parent","degree"),
   functions.name="functions",
   track.height = 2.5,
   track.margin = c(0,0),
   label.cex=0.6,
   wrap.at = " ",
   facing=c("inside","inside","inside"),parsefun=parser)
   \dontrun{
   plotChordDiagram(
     rv144_aggregate,
     threshold = 0.85,
     varnames = c("functions"),
     separator = " ",
     function.separator = ",",
     track.order = "functions",
     sort.order = c("degree", "functions"),
     inner.track = "functions",
     parsefun = parser,
     track.height = 2,
     track.margin = c(0, 0),
     functions.name = c("functions"),
     facing = "inside",
     wrap.at = " ",
     label.cex = 0.6
)
   }
}
