% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{print.flowReMix}
\alias{print.flowReMix}
\title{print.flowReMix}
\usage{
\method{print}{flowReMix}(x, ...)
}
\arguments{
\item{x}{a flowReMix fit}

\item{...}{}
}
\description{
print a flowReMix object
}
