% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityFunctions.R
\name{stabilityGraph}
\alias{stabilityGraph}
\title{compute a stability graph for the Ising model from a flowReMix fit.}
\usage{
stabilityGraph(obj, type = c("ising", "randomEffects"), cv = FALSE,
  reps = 100, cpus = 1, gamma = 0.9, AND = TRUE)
}
