% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reMixMisc.R
\name{stimulationModel}
\alias{stimulationModel}
\title{Preparing a Dataset for a Stimulation Respone Analysis with flowReMix}
\usage{
stimulationModel(data, cell_type, stim_var, controls, stim_groups)
}
\arguments{
\item{data}{the data set to be transformed.}

\item{cell_type}{a factor vector identifying which cell type each row in the data
refers to.}

\item{stim_var}{a vector identifying the which stimulation each row in the data refers to.}

\item{controls}{a vector identifying which of the stimulations are controls.}

\item{stim_groups}{a list detailing which stimulations are to be grouped together.}
}
\description{
Given a dataset and a list of stimulation groups, the function pairs each
stimulation group with control stimulations and defines a new cell_type variable
which is the interaction of the previous cell type with stimulation group.
}
