% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summary}
\alias{summary}
\alias{summary.flowReMix}
\title{summary of a flowReMix fit}
\usage{
\method{summary}{flowReMix}(object, ...)
}
\arguments{
\item{object}{a flowReMix model fit.}

\item{...}{additional arguments. See description}
}
\description{
summarize the output of a flowReMix object into a rocTable
Uses non-standard evaluation
\itemize{
\item{subject}{The name of the subject variable, as an unquoted variable.}
\item{target}{The name of the outcome variable as an unquoted variable.}
\item{type}{either "ROC" or "FDR".}
}
}
