% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translateCOMPASStoflowReMixNames}
\alias{translateCOMPASStoflowReMixNames}
\title{Translate COMPASS markere names to FlowReMix format}
\usage{
translateCOMPASStoflowReMixNames(x)
}
\arguments{
\item{x}{\code{matrix} with column names in COMPASS format}
}
\value{
\code{matrix} with column names in FlowReMix format.
}
\description{
Translate COMPASS markere names to FlowReMix format
}
\examples{
\dontrun{
#x is a matrix with columns named by COMPASS.
translateCOMPASStoFlowReMixNames(x)
}
}
