% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weightForPFS}
\alias{weightForPFS}
\title{weightForPFS
Compute the weighting for the polyfunctionality score for each type of subset.}
\usage{
weightForPFS(x, M = NULL, parser = degreeFromStringFun)
}
\arguments{
\item{x}{\code{flowReMix} fit object}

\item{M}{\code{numeric} the max possible functionality. Uses \code{max(degree)} if not provided.}

\item{parser}{\code{function} to parse the degree from a cell subset name. Defaults to \code{\link{degreeFromStringFun}} in the package.}
}
\value{
\code{numeric} vector of weights.
}
\description{
weightForPFS
Compute the weighting for the polyfunctionality score for each type of subset.
}
\details{
Computes the weight for the polyfunctionality score calculation. The weight is \code{choose(M,k)} where
k is the degree of a subset and M is the \code{max(degree)} or the number of markers.
}
\examples{
data(fit505)
weightForPFS(fit505)
}
\seealso{
\code{\link{flowReMixPFS}} \code{\link{degreeFromStringFun}}
}
