% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeroPosteriorProbs}
\alias{zeroPosteriorProbs}
\title{zeroPosteriorProbs}
\usage{
zeroPosteriorProbs(modelfit)
}
\arguments{
\item{modelfit}{A flowReMix model fit}
}
\value{
matrix of posterior probabilities
}
\description{
return the posterior probabilities with preassignments set to zero.
}
