% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glmerMixture.R
\name{glmmMixture}
\alias{glmmMixture}
\title{Estimate a composite mixture GEE}
\usage{
glmmMixture(formula, sub.population = NULL, N = NULL, id, waves,
  data = parent.frame(), treatment, treatment.levels = 2,
  compute.offset = TRUE, weights = NULL, init.beta = NULL, init.rho = 0,
  rho.fixed = FALSE, shrinkage.folds = 5, maxiter = 100, tol = 0.001,
  nAGQ = 1, computeT = FALSE, shrink.sandwich = TRUE)
}
\description{
Estimate a composite mixture GEE
}

