% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{plot}
\alias{plot}
\alias{plot.flowReMix}
\title{plot a flowReMix fit object}
\usage{
\method{plot}{flowReMix}(x, ...)
}
\arguments{
\item{...}{additional arguments.}

\item{obj}{The model fit of class \code{flowReMix} returned by the fitting function.}

\item{type}{the type of plot to make, one of c("scatter","boxplot","FDR","ROC","graph")}

\item{target}{the name of the outcome variable as an unquoted variable.}

\item{varname}{the variable name to appear in the legend}
}
\description{
Generate a plot of a flowReMix model fit. Various plot types are supported.
"scatter", "boxplot", "FDR", "ROC", "ising"
}
