% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{plot}
\alias{plot}
\alias{plot.flowReMix}
\title{plot a flowReMix fit object}
\usage{
\method{plot}{flowReMix}(x,...)
}
\arguments{
\item{x}{The model fit of class \code{flowReMix} returned by the fitting function.}

\item{...}{additional arguments, see description.}
}
\description{
Generate a plot of a flowReMix model fit. Various plot types are supported.
\itemize{
\item{subject}{The name of the subject variable, as an unquoted variable.}
\item{target}{unquoted variable name in the data corresponding to the outcome variable. Required.}
\item{varname}{character string used to label the outcome variable on the plot legend.}
\item{type}{character specifying the plot type. can be "scatter", "boxplot", "FDR", "ROC", "ising". Required.}
}
}
\examples{
library(ggplot2)
data(fit505)
plot(fit505,target=vaccine,type="scatter",subsets=getSubsets(fit505)[1:4])+
     facet_wrap(~sub.population,ncol=2)
plot(fit505,target=vaccine,type = "boxplot", jitter=TRUE,test="none",
     group = list(getSubsets(fit505)[1:4]),varname="vaccination")
plot(getIsing(fit505),threshold=0.9,
     fill=summary(fit505,target=vaccine)$auc)
plot(fit505,target=vaccine,type = "ROC",
     subsets = getSubsets(fit505)[3:4],varname="vaccination")
plot(fit505,target=vaccine,type = "FDR",
     subsets = getSubsets(fit505)[3:4],varname="vaccination")


}
