% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityFunctions.R
\name{stabilityGraph}
\alias{stabilityGraph}
\title{compute a stability graph for the Ising model from a flowReMix fit.}
\usage{
stabilityGraph(obj, type = c("ising", "randomEffects"), cv = FALSE,
  reps = 100, cpus = 1, gamma = 0.9, AND = TRUE, seed = NULL,
  sampleNew = FALSE, keepEach = 1)
}
\arguments{
\item{obj}{\code{flowReMix} model fit.}

\item{type}{\code{character} "ising" or "randomEffects"}

\item{cv}{\code{integer} MORE DETAILS}

\item{reps}{\code{integer} number of replications.}

\item{cpus}{\code{integer} number of cpus to use}

\item{gamma}{\code{numeric} MORE DETAILS}

\item{AND}{\code{logical}. How the Ising edges are aggregated. AND or OR.}

\item{seed}{random seed. Default 100 if NULL.}

\item{sampleNew}{\code{logical} Do we generate new samples or use the ones from the fitting procedure? (default FALSE)}

\item{keepEach}{\code{numeric} Keep every \code{keepEach} simulation result. (default 1)}
}
\description{
Compute the stable components of the Ising or Random Effects graphical model through bootstrapping.
}
