% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFunctions.R
\name{summary}
\alias{summary}
\alias{summary.flowReMix}
\title{summary of a flowReMix fit}
\usage{
\method{summary}{flowReMix}(object, ...)
}
\arguments{
\item{subject}{the name of the subject variable, as an unquoted variable}

\item{target}{the name of the outcome variable as an unquoted variable.}

\item{type}{either "ROC" or "FDR".}
}
\description{
summarize the output of a flowReMix object into a rocTable
Uses non-standard evaluation
}
